Add script for wordcloud generation
