+++ app/myapp/templates/myapp/yearbook.html
@@ -111,7 +111,7 @@
 						{{student.future}}
 					</div>				
 				</div>
-					{% if student.AdjectivesIGet.count %}
+					{% if student.WordCloud %}
 						<div class="upperContRightC2">
 							<img style="height:200px;width: auto;max-width: 100%;" src="{{student.WordCloud.url}}" alt="" />
 						</div>
